Move a bunch of crate stuff to CrateExtensions
- CrateExtensions is a static class which groups extra stuff that the crate systems use.
