+++ Project/Assets/_Data/Scripts/Crates/CrateExtensions.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: f9f4d875ede6d0b448a79dc4f4bcab0c
