+++ Project/Assets/_Data/Scripts/Crates/CrateObject.cs
@@ -1,6 +1,5 @@
-using System;
 using UnityEngine;
-
+using static CrateExtensions;
 
 /// <summary>
 ///  This class is mostly for demonstration.
@@ -41,33 +40,12 @@ public class CrateObject : MonoBehaviour, ICollectable
             // Colour this object based on its tag
             var renderer = GetComponent<Renderer>();
             renderer.GetPropertyBlock(block);
-            block.SetColor("_BaseColor", GetColourFromTag(value));
+            block.SetColor("_BaseColor", value.GetColourFromTag());
             renderer.SetPropertyBlock(block);
         } 
     }
 
     public GameObject GameObject { get => gameObject; }
  
-    // Types of tags a crate can have. Add more to the enum if you want.
-    // This can be referenced by calling CrateObject.CrateTag. 
-    public enum CrateTag { Red, Green, Blue }
-
-    // Gets a random crate tag
-    public static CrateTag GetRandomCrateTag()
-    {
-        int length = Enum.GetNames(typeof(CrateTag)).Length;
-        return (CrateTag)UnityEngine.Random.Range(0, length);
-    }
-
-    // Returns a colour for a given colour-named tag
-    public static Color GetColourFromTag(CrateTag tag)
-    {
-        return tag switch
-        {
-            CrateTag.Red => Color.red,
-            CrateTag.Green => Color.green,
-            CrateTag.Blue => Color.blue,
-            _ => Color.white,
-        };
-    }
+    
 }
