+++ Project/Assets/_Data/Scripts/Interfaces/ICollectable.cs
@@ -8,6 +8,6 @@ public interface ICollectable
 {
     float Score { get; set; }
     GameObject GameObject { get; }
-    CrateObject.CrateTag Tag { get; set; }
+    CrateExtensions.CrateTag Tag { get; set; }
 
 }
